Create RemoteMachineFinder
