+++ config/services_test.yaml
@@ -3,6 +3,12 @@ services:
         autowire: true
         autoconfigure: true
 
+    App\Services\RemoteMachineFinder:
+        public: true
+        arguments:
+            $machineManagers:
+                - '@App\Services\DigitalOceanMachineManager'
+
     Symfony\Component\Messenger\Transport\InMemoryTransport:
         alias: 'messenger.transport.async'
